_getDisplay() should return primary display
This avoids issues in Windows when working with multiple monitors. If you start the app in a desktop and then move your mouse accidentally to the other desktop windows appears out of place. Also it makes sense to work with the primary display as it contains the taskbar app icons.